1、完成导出保险清单！

+++ src/main/java/cn/edu/guet/insuranceandmove/service/InsuranceListService.java
@@ -23,6 +23,6 @@ public interface InsuranceListService extends IService<InsuranceList> {
 
     List<InsuranceStatistics> selectInsuranceStatisticsByYear(int year);
 
-    void simpleWrite(List<Long> idsList);
+    void simpleWrite(List<Integer> idsList);
 
 }
